requirements.txt
python-telegram-bot==13.15
yfinance
pandas
urllib3==1.26.16
